==> pipeline/notebooks/1_translate_title.ipynb <==
# 1: Translate title from French to English

Since the phrase grounding models selected were trained on English captions, the titles of the images need to be translated from French to English. A pretrained machine learning model, [MarianMT](https://huggingface.co/Helsinki-NLP/opus-mt-fr-en), which is a multilingual machine translation model trained on the on the [OPUS](https://opus.nlpl.eu/) corpus, is used for this purpose.

import pickle
import sqlite3
import pandas as pd
from transformers import AutoTokenizer, AutoModelForSeq2SeqLM

We'll be using Helsinki-NLP's [opus-mt](https://huggingface.co/Helsinki-NLP) model to translate the titles from French to English.

tokenizer = AutoTokenizer.from_pretrained("Helsinki-NLP/opus-mt-fr-en")

model = AutoModelForSeq2SeqLM.from_pretrained("Helsinki-NLP/opus-mt-fr-en")

model = model.cuda()

# Read sqlite query results into a pandas DataFrame
con = sqlite3.connect("../data/BCU_database/bcu_caption.db")
Photo = pd.read_sql_query("SELECT * from Photo", con) 
con.close()

# Verify that result of SQL query is stored in the dataframe
Photo.head()

def translate(sentence):
    input_ids = tokenizer(sentence, return_tensors="pt").input_ids.cuda()
    outputs = model.generate(input_ids=input_ids, num_beams=5, num_return_sequences=1)
    return tokenizer.batch_decode(outputs, skip_special_tokens=True)[0]

Photo['title_en'] = Photo['title'].apply(lambda x: translate(x) if pd.notna(x) else '')

Photo.to_sql("Photo",con,if_exists='replace',index=False, index_label="id")
