==> pipeline/notebooks/2_expression_preprocessing.ipynb <==
# 2: Expression preprocessing

We first link the *bcu_id* to the filenames. 

Then the captions and titles are preprocessed in preparation for phrase grounding. This is achieved by converting the text to lowercase and removing expressions such as "a picture of", "a view of" and "a photograph of". As this dataset consists of images from Fribourg, there are many mentions of Fribourg in the title, which can potentially confuse the phrase grounding models. Therefore, it was decided to remove these mentions.

import glob
import sqlite3
from os import listdir
from os.path import isfile, join

import numpy as np
import pandas as pd

BCU_database_folder = '../data/BCU_database/'

Extract the Photo table from the database

# Read sqlite query results into a pandas DataFrame
con = sqlite3.connect(BCU_database_folder + 'bcu_caption.db')
Photo = pd.read_sql_query("SELECT * from Photo", con)

# Verify that result of SQL query is stored in the dataframe
print('original number of picture :',len(Photo))
Photo.head()

## Match bcu_id to filenames

mypath = BCU_database_folder + '03_resized'
onlyfiles = [f for f in listdir() if isfile(join(mypath, f))]

files = glob.glob(mypath + '/*')

files_324w = list(filter(lambda k: '324w' in k, files))

missing = []
Photo['filename'] = Photo.loc[~Photo.bcu_id.str.contains('BCUF')].bcu_id.apply(lambda x:  (list(filter(lambda k: x.lower() in k.lower(), files_324w)))[0].removeprefix(mypath + '/') if len(list(filter(lambda k: x.lower() in k.lower(), files_324w))) > 0 else missing.append(x))
print(f'Number of pictures with filenames: {Photo.filename.notnull().sum()}')

Processing titles and captions

Photo['title_en_preprocessed'] = Photo["title_en"].apply(lambda x: np.nan if str(x).find('Souvenir of') != -1 else str(x).lower().replace('portrait of ', '').replace('canton of fribourg','').replace('(fribourg)','').replace('[fribourg]','').replace('of fribourg','').replace('fribourg',''))
Photo['caption_preprocessed'] = Photo["caption"].apply(lambda x: np.nan if (str(x).find('a city in') != -1 or str(x).find('view') != -1) else str(x).replace('portrait of ', '').replace('photograph of ', '').replace('black and white photo of ', ''))
print(f'Number of images with processed captions and filenames: {(Photo.caption_preprocessed.notnull() & Photo.filename.notnull()).sum()}')
print(f'Number of images with processed titles and filenames: {(Photo.title_en_preprocessed.notnull() & Photo.filename.notnull()).sum()}')


Photo.to_sql("Photo",con,if_exists='replace',index=False, index_label="id")
