==> pipeline/notebooks/4_MDETR_inference_colab.ipynb <==
# MDETR - Modulated Detection for End-to-End Multi-Modal Understanding

Welcome to the demo notebook for MDETR. We'll show-case detection, segmentation and question answering



## Preliminaries

This section contains the initial boilerplate. Run it first.

!pip install pandas --upgrade

from google.colab import drive
drive.mount('/content/drive')

!pip install timm==0.4.5 transformers==4.5.1

!cp /content/drive/MyDrive/BCUFR/dataset_for_phrase_grounding/dataset_for_phrase_grounding.zip dataset.zip &&  unzip -q dataset.zip 
!cp /content/drive/MyDrive/BCUFR/4_all_inferences.pkl .

!wget https://raw.githubusercontent.com/tgieruc/Heritage-in-the-digital-age/main/pipeline/4_MDETR_inference.py

!python 4_MDETR_inference.py --input_file 2_preprocessing.pkl --output_file temp.pkl --expression_column caption_preprocessed --inference_column MDETR_caption --image_directory img


!python 4_MDETR_inference.py --input_file temp.pkl --output_file 4_mdetr_caption_title.pkl --expression_column title_en_preprocessed --inference_column MDETR_title --image_directory img


!cp 4_mdetr_caption_title.pkl /content/drive/MyDrive/BCUFR/4_mdetr_caption_title.pkl
