==> pipeline/notebooks/5_label_postprocess.ipynb <==
# 5: Caption postprocessing
The phrase grounding models tend to use the word "group" as a label, rather than the noun following it. This error is fixed by checking if the word "group" is followed by a noun, and if it is, the label is changed to the noun. 


import pickle
import sys
from pathlib import Path

import pandas as pd

import src.GLIP.maskrcnn_benchmark as maskrcnn_benchmark

sys.modules['maskrcnn_benchmark'] = maskrcnn_benchmark


phrase_grounding_selected_folder = '../data/phrase_grounding_selected/'
dataset_for_phrase_grounding = '../data/dataset_for_phrase_grounding/'


# iterate over files in
# that directory
files = Path(phrase_grounding_selected_folder).glob('*.p')

pickle_filename = []
for file in files:
    pickle_filename.append(str(file))
pickle_filename

dataset = pd.DataFrame()
for pickle_file in pickle_filename:
    dataset = dataset.append(pickle.load(open(pickle_file, 'rb')), ignore_index=True)

dataset

dataset['conf'] = dataset.results.apply(lambda x: x[0])
dataset['bbox'] = dataset.results.apply(lambda x: x[1])
dataset['labels'] = dataset.results.apply(lambda x: x[2])
dataset = dataset.drop(columns='results')

expr_dataset = pickle.load(open(dataset_for_phrase_grounding + 'dataset.p', 'rb'))
expr_dataset = pd.DataFrame(expr_dataset)
expr_dataset['caption_raw'] = expr_dataset.caption.apply(lambda x: x['raw'])
expr_dataset['caption_preprocessed'] = expr_dataset.caption.apply(lambda x: x['preprocessed'])
expr_dataset['title_raw'] = expr_dataset.title.apply(lambda x: x['raw'])
expr_dataset['title_preprocessed'] = expr_dataset.title.apply(lambda x: x['preprocessed'])
expr_dataset = expr_dataset.drop(columns=['caption', 'title'])
expr_dataset.head()

dataset = pd.merge(dataset, expr_dataset, left_on='filename', right_on='filename', how='left')

def remove_group(elem):
    if ('a group' in elem['labels']) or ('group' in elem['labels']):
        if elem['expr'] == 'caption':
            expression = elem['caption_preprocessed']
        else:
            expression = elem['title_preprocessed']
        expression = expression.split(' ')
        try:
            word = expression[expression.index('of') + 1]
            if word == 'young':
                word = word + ' ' + expression[expression.index('of') + 2]
            return [word if label == 'a group' else label for label in elem['labels']]
        except:
            return elem['labels']

    else:
        return elem['labels']

dataset['labels'] = dataset.apply(lambda x: remove_group(x), axis=1)

pickle.dump(dataset, open('../data/dataset_for_segmentation.p','wb'))
